shopping_mall/backend/pyproject.toml
[project]
name = "shopping-mall-backend"
version = "0.1.0"
description = "Shopping Mall Dummy API Server (FastAPI + PostgreSQL)"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy>=2.0.36",
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
    "apscheduler>=3.10.0",
    "chromadb>=0.5.0",
    "python-jose[cryptography]>=3.3.0",
    "ollama>=0.6.1",
    "anthropic>=0.40.0",
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
]

[tool.uv]
exclude-newer = "2026-03-26T00:00:00Z"
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
